@@ -1,9 +1,18 @@
 module scribe-backend
 
-go 1.20
+go 1.21
 
 require (
+	github.com/didip/tollbooth/v7 v7.0.1
 	github.com/google/uuid v1.3.0
 	github.com/gorilla/mux v1.8.0
 	github.com/gorilla/websocket v1.5.0
+	github.com/rs/zerolog v1.30.0
+)
+
+require (
+	github.com/go-pkgz/expirable-cache v0.1.0 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.19 // indirect
+	golang.org/x/sys v0.11.0 // indirect
 )
